+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- Validate WinSW configuration on install/start/restart
+- Ability to configure extensions and more complex logging modes
+- Default to WinSW 2.x binary
+
 # 0.1.0
 
-Initial release of winsw
+- Initial release of WinSW cookbook